requirements.txt
gunicorn==23.0.0
asyncpg==0.30.0
django==5.1.6
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
numpy==2.2.3
pillow==11.1.0
opencv-python==4.11.0.86
requests==2.32.3